Fixed a bug in determining the starting player.
Before the patch: If the game was not started, pressing new game does 
not switch starting player. In the middle of the game it does switch. 
But if the game is ended, it doesn't.

After the patch: The table is clear after starting new game. I call 
invalidate() on the scene.

Patch provided by: Pelladi Gabor
CCMAIL:pelladigabor@gmail.com

svn path=/trunk/KDE/kdegames/bovo/; revision=943519