Fixed a bug in determining the starting player.
Before the patch: If the game was not started, pressing new game does 
not switch starting player. In the middle of the game it does switch. 
But if the game is ended, it doesn't.

After the patch: When starting a new game after a finished one, the 
starting player switches. Other cases remain the same.

Patch provided by: Pelladi Gabor
CCMAIL:pelladigabor@gmail.com

svn path=/trunk/KDE/kdegames/bovo/; revision=943522